Read operations now have access to a local cache object
read() will look for cached values for the requested
section and return those if available rather than
performing a network request.

Calls to write() will automatically clear the cache
for that section.

Use hardRead() with the same arguments as read() to
clear the cache for thew requested section and
force a network request.